#122 - Minor refactoring
